# fix: try to load class from multiple class-loaders