Basically, we borrowed `ClassUtils` from Spring, which prioritizes `Thread::getContextClassLoader` over `ClassUtils.class.getClassLoader()`. It is fine in most situations. But sometimes, the library is loaded with an isolated classloader, yet provided with a system classloader in `Thread::getContextClassLoader`, which causes problems.

We simply retry with `ClassUtils.class.getClassLoader()` when failing to load a class. It is not the most efficient, but should be enough (if people cache the lookup results):

```lua
String = java.import('java.lang.String')
s = String('s')
-- instead of
s = java.import('java.lang.String')('s')
```

Also, I also think it is a good chance that we optimize our reflection usage. We were not caching the method lookup results (because Lua is dynamic, and we *do* need to check if every parameter matches (and convert them over) every single time.) Caching the results improves a bit for `L.run("assert(big_num:intValue() == 1024")` but we do need better profiling to get anything meaningful:

```
--- before
Benchmark                                             (lua)  Mode  Cnt  Score   Error  Units
MethodCallBenchmark.benchmarkObjectMethodCall       Lua 5.4  avgt    3  4.660 ± 0.234  us/op
MethodCallBenchmark.benchmarkObjectMethodCall        LuaJIT  avgt    3  3.932 ± 0.100  us/op
MethodCallBenchmark.benchmarkObjectMethodCall          LuaJ  avgt    3  4.003 ± 0.246  us/op
--- after
Benchmark                                             (lua)  Mode  Cnt  Score   Error  Units
MethodCallBenchmark.benchmarkObjectMethodCall       Lua 5.4  avgt    3  2.300 ± 0.038  us/op
MethodCallBenchmark.benchmarkObjectMethodCall        LuaJIT  avgt    3  1.698 ± 0.237  us/op
MethodCallBenchmark.benchmarkObjectMethodCall          LuaJ  avgt    3  1.673 ± 0.018  us/op
```

fixes #200